Generic support for Swift TempUrl middleware
