+++ apis/cloudfiles/src/main/java/org/jclouds/cloudfiles/blobstore/config/CloudFilesBlobStoreContextModule.java
@@ -25,14 +25,14 @@ import javax.inject.Singleton;
 
 import org.jclouds.blobstore.BlobRequestSigner;
 import org.jclouds.cloudfiles.CloudFilesClient;
-import org.jclouds.cloudfiles.TemporaryUrlKey;
+import org.jclouds.openstack.swift.TemporaryUrlKey;
 import org.jclouds.cloudfiles.blobstore.CloudFilesAsyncBlobStore;
-import org.jclouds.cloudfiles.blobstore.CloudFilesBlobRequestSigner;
 import org.jclouds.cloudfiles.blobstore.CloudFilesBlobStore;
 import org.jclouds.cloudfiles.blobstore.functions.CloudFilesObjectToBlobMetadata;
 import org.jclouds.cloudfiles.domain.ContainerCDNMetadata;
 import org.jclouds.date.TimeStamp;
 import org.jclouds.openstack.swift.blobstore.SwiftAsyncBlobStore;
+import org.jclouds.openstack.swift.blobstore.SwiftBlobRequestSigner;
 import org.jclouds.openstack.swift.blobstore.SwiftBlobStore;
 import org.jclouds.openstack.swift.blobstore.config.SwiftBlobStoreContextModule;
 import org.jclouds.openstack.swift.blobstore.functions.ObjectToBlobMetadata;
@@ -64,18 +64,6 @@ public class CloudFilesBlobStoreContextModule extends SwiftBlobStoreContextModul
       });
    }
 
-   @Provides
-   @TimeStamp
-   protected Long unixEpochTimestampProvider() {
-      return System.currentTimeMillis() / 1000; /* convert to seconds */
-   }
-
-   @Provides
-   @TemporaryUrlKey
-   protected String temporaryUrlKeyProvider(CloudFilesClient client) {
-      return client.getTemporaryUrlKey();
-   }
-
    @Override
    protected void configure() {
       super.configure();
@@ -86,6 +74,6 @@ public class CloudFilesBlobStoreContextModule extends SwiftBlobStoreContextModul
 
    @Override
    protected void configureRequestSigner() {
-      bind(BlobRequestSigner.class).to(CloudFilesBlobRequestSigner.class);
+      bind(BlobRequestSigner.class).to(SwiftBlobRequestSigner.class);
    }
 }
