+++ apis/swift/src/main/java/org/jclouds/openstack/swift/functions/ParseTemporaryUrlKeyFromHeaders.java
@@ -16,13 +16,15 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.jclouds.cloudfiles.functions;
+package org.jclouds.openstack.swift.functions;
 
 import com.google.common.base.Function;
+import com.google.common.collect.Multimap;
 import org.jclouds.http.HttpResponse;
+import org.jclouds.openstack.swift.reference.SwiftHeaders;
 
 import static com.google.common.collect.Iterables.getOnlyElement;
-import static org.jclouds.cloudfiles.reference.CloudFilesHeaders.ACCOUNT_TEMPORARY_URL_KEY;
+import static org.jclouds.openstack.swift.reference.SwiftHeaders.ACCOUNT_TEMPORARY_URL_KEY;
 
 /**
  * @author Andrei Savu
@@ -31,6 +33,11 @@ public class ParseTemporaryUrlKeyFromHeaders implements Function<HttpResponse, S
 
    @Override
    public String apply(HttpResponse httpResponse) {
-      return getOnlyElement(httpResponse.getHeaders().get(ACCOUNT_TEMPORARY_URL_KEY));
+      Multimap<String, String> headers = httpResponse.getHeaders();
+      if (headers.containsKey(ACCOUNT_TEMPORARY_URL_KEY)) {
+         return getOnlyElement(headers.get(ACCOUNT_TEMPORARY_URL_KEY));
+      } else {
+         return null;
+      }
    }
 }
