+++ apis/swift/src/test/java/org/jclouds/openstack/swift/internal/StubSwiftAsyncClient.java
@@ -31,6 +31,7 @@ import java.util.concurrent.ExecutorService;
 import javax.inject.Inject;
 import javax.inject.Named;
 import javax.inject.Singleton;
+import javax.ws.rs.HeaderParam;
 
 import org.jclouds.Constants;
 import org.jclouds.blobstore.LocalAsyncBlobStore;
@@ -60,6 +61,7 @@ import com.google.common.base.Throwables;
 import com.google.common.collect.Iterables;
 import com.google.common.collect.Sets;
 import com.google.common.util.concurrent.ListenableFuture;
+import org.jclouds.openstack.swift.reference.SwiftHeaders;
 
 /**
  * Implementation of {@link SwiftAsyncClient} which keeps all data in a local Map object.
@@ -201,7 +203,17 @@ public class StubSwiftAsyncClient implements CommonSwiftAsyncClient {
         return null;
     }
 
-    public ListenableFuture<Boolean> setObjectInfo(String container, String key, Map<String, String> userMetadata) {
+   @Override
+   public ListenableFuture<String> getTemporaryUrlKey() {
+      throw new UnsupportedOperationException();
+   }
+
+   @Override
+   public ListenableFuture<Void> setTemporaryUrlKey(String key) {
+      throw new UnsupportedOperationException();
+   }
+
+   public ListenableFuture<Boolean> setObjectInfo(String container, String key, Map<String, String> userMetadata) {
       throw new UnsupportedOperationException();
    }
